# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Cat F\u00fcd API",
        "version": "0.0.1",
        "description": "The Cat F\u00fcd API: a shared pet-food catalogue (brand ladder, products, barcodes, advisories) plus each user's private pets, pantry, ratings and meal history.\n\nAuthenticate with `POST /register` or `POST /login`, then send `Authorization: Bearer <token>`. Catalogue reads are scoped to the caller's region and\nto what they may see: approved products, plus their own pending contributions. Staff (moderator, admin) see more fields."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/brand-nodes/{brandNode}": {
            "patch": {
                "operationId": "product.updateNode",
                "summary": "Rename, re-kind, re-species or move a brand-tree node. A rename keeps the old name as an alias",
                "tags": [
                    "Product"
                ],
                "parameters": [
                    {
                        "name": "brandNode",
                        "in": "path",
                        "required": true,
                        "description": "The brand node ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBrandNodeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "node": {
                                            "$ref": "#/components/schemas/BrandNodeResource"
                                        }
                                    },
                                    "required": [
                                        "node"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/brand-nodes/{brandNode}/merge": {
            "post": {
                "operationId": "product.mergeNode",
                "summary": "Fold one node into another. Products, lines and barcodes move across; the old names become aliases",
                "tags": [
                    "Product"
                ],
                "parameters": [
                    {
                        "name": "brandNode",
                        "in": "path",
                        "required": true,
                        "description": "The brand node ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MergeBrandNodeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "node": {
                                            "$ref": "#/components/schemas/BrandNodeResource"
                                        }
                                    },
                                    "required": [
                                        "node"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/products/audit-summary": {
            "get": {
                "operationId": "product.auditSummary",
                "summary": "Progress for the seed -> wire-up -> audit pass",
                "tags": [
                    "Product"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "total": {
                                            "type": "integer"
                                        },
                                        "without_barcode": {
                                            "type": "integer"
                                        },
                                        "without_image": {
                                            "type": "integer"
                                        },
                                        "unreviewed": {
                                            "type": "integer"
                                        },
                                        "reviewed": {
                                            "type": "integer"
                                        },
                                        "needs_changes": {
                                            "type": "integer"
                                        }
                                    },
                                    "required": [
                                        "total",
                                        "without_barcode",
                                        "without_image",
                                        "unreviewed",
                                        "reviewed",
                                        "needs_changes"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/products/{product}/versions": {
            "get": {
                "operationId": "product.versions",
                "summary": "Every recipe/label on record for a product, newest first",
                "tags": [
                    "Product"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "versions": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductVersionResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "versions"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "UpdateBrandNodeRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Renaming keeps the old name as an alias.",
                        "maxLength": 255
                    },
                    "parent_id": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "description": "Move under another node; `null` makes it a top-level node."
                    },
                    "kind": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "enum": [
                            "manufacturer",
                            "brand",
                            "line",
                            "subline",
                            null
                        ]
                    },
                    "aliases": {
                        "type": [
                            "array",
                            "null"
                        ],
                        "items": {
                            "type": "string",
                            "maxLength": 255
                        },
                        "maxItems": 30
                    },
                    "species": {
                        "type": [
                            "array",
                            "null"
                        ],
                        "description": "Species this is sold for (`cat`, `dog`); `null` or empty means all.",
                        "items": {
                            "type": "string",
                            "maxLength": 30
                        },
                        "maxItems": 10
                    },
                    "notes": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 5000
                    }
                },
                "title": "UpdateBrandNodeRequest"
            },
...
        },
        "responses": {
...
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}